tests: Migrate ctl_v3_watch_no_cov_test to ctl_v3_watch_test No need to separate files.
